Make logging optional and use decimal for prices
Make logger parameters optional across services and controllers, defaulting to NullLogger when not provided. This improves testability and reduces DI coupling.

Change price fields (Cost, DisplayPrice, DiscountedPrice) from double to decimal in DTOs for better precision with monetary values.

+++ POS_System/Models/Dto/ProductLineItemDto.cs
@@ -5,9 +5,9 @@
         public string Id { get; set; } = string.Empty;
         public string BarCodeId { get; set; } = string.Empty;
         public string ProductId { get; set; } = string.Empty;
-        public double Cost { get; set; }
-        public double DisplayPrice { get; set; }
-        public double DiscountedPrice { get; set; }
+        public decimal Cost { get; set; }
+        public decimal DisplayPrice { get; set; }
+        public decimal DiscountedPrice { get; set; }
         public int Quantity { get; set; }
